pyproject.toml
[project]
name = "poetry"
version = "0.1.0"
description = ""
authors = [
    {name = "buddyuchiha",email = "Pro100biv@mail.ru"}
]
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi (>=0.128.0,<0.129.0)",
    "uvicorn (>=0.40.0,<0.41.0)",
    "sqlalchemy (>=2.0.46,<3.0.0)",
    "alembic (>=1.18.3,<2.0.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "psycopg2 (>=2.9.11,<3.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "bcrypt (==4.0.1)",
    "passlib (>=1.7.4,<2.0.0)",
]

[tool.poetry]
packages = [{include = "poetry", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
